Improve handling of KeycloakClient CRs and client secrets.
This ensures that the KeycloakClient secret is not regenerated on each
reconcilation loop. We also no longer generate a password and instead
allow the Keycloak operator to do so; it generates a random UUID
and stores it in a Secret that we can fetch.

This means that Gitea config no longer has a simple means to identify when
the client-secret has changed and issue the update-oauth commands, and so
with this change that capability is removed. This should probably end up
being a function of the Gitea operator (e.g. watching for last updated
times of the corresponding Secret).